+++ server/jetstream_test.go
@@ -11,6 +11,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build !skip_js_tests
+// +build !skip_js_tests
+
 package server
 
 import (
@@ -317,7 +320,7 @@ func TestJetStreamAutoTuneFSConfig(t *testing.T) {
 	s := RunRandClientPortServer()
 	defer s.Shutdown()
 
-	jsconfig := &JetStreamConfig{MaxMemory: -1, MaxStore: 128 * 1024 * 1024 * 1024}
+	jsconfig := &JetStreamConfig{MaxMemory: -1, MaxStore: 128 * 1024 * 1024}
 	if err := s.EnableJetStream(jsconfig); err != nil {
 		t.Fatalf("Expected no error, got %v", err)
 	}
@@ -361,7 +364,7 @@ func TestJetStreamAutoTuneFSConfig(t *testing.T) {
 	testBlkSize("foo", 1, 512, FileStoreMinBlkSize)
 	testBlkSize("foo", 1, 1024*1024, defaultMediumBlockSize)
 	testBlkSize("foo", 1, 8*1024*1024, defaultMediumBlockSize)
-	testBlkSize("foo_bar_baz", -1, 32*1024*1024*1024, FileStoreMaxBlkSize)
+	testBlkSize("foo_bar_baz", -1, 32*1024*1024, FileStoreMaxBlkSize)
 }
 
 func TestJetStreamConsumerAndStreamDescriptions(t *testing.T) {
