+++ test/fanout_test.go
@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !race
-// +build !race
+//go:build !race && !skipnoracetests
+// +build !race,!skipnoracetests
 
 package test
